refactor: reorganize report packages into summary, renderer hierarchy
Move the summary builder to study/report/summary/ and the terminal
renderer to study/report/renderer/terminal/. The core report package
retains only types, interfaces, and Section conversion functions.

Package layout:
  study/report/              -- types, Section interface, conversion helpers
  study/report/summary/      -- Build() and portfolio-specific builders
  study/report/renderer/terminal/ -- lipgloss + ntcharts styled output
