+++ cli/backtest.go
@@ -13,8 +13,8 @@ import (
 	"github.com/penny-vault/pvbt/data"
 	"github.com/penny-vault/pvbt/engine"
 	"github.com/penny-vault/pvbt/portfolio"
-	backtestReport "github.com/penny-vault/pvbt/study/report"
-	"github.com/penny-vault/pvbt/study/report/terminal"
+	"github.com/penny-vault/pvbt/study/report/renderer/terminal"
+	"github.com/penny-vault/pvbt/study/report/summary"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
@@ -194,7 +194,7 @@ func runBacktest(cmd *cobra.Command, strategy engine.Strategy) error {
 
 	log.Info().Str("path", outputPath).Msg("backtest output written")
 
-	rpt, err := backtestReport.Summary(acct)
+	rpt, err := summary.Build(acct)
 	if err != nil {
 		log.Warn().Err(err).Msg("some report metrics failed")
 	}
