+++ study/report/renderer/terminal/chart.go
