+++ study/report/renderer/terminal/drawdowns.go
