+++ study/report/renderer/terminal/header.go
