+++ study/report/renderer/terminal/monthly.go
