+++ study/report/renderer/terminal/renderer.go
