+++ study/report/renderer/terminal/returns.go
