+++ study/report/renderer/terminal/risk.go
