+++ study/report/renderer/terminal/style.go
