+++ study/report/renderer/terminal/trades.go
